requirements.txt
# Google Cloud dependencies
google-cloud-firestore==2.13.1
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.16.4

# Web framework
flask==3.0.0
flask-cors==4.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# API and utilities
requests==2.31.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Production server
gunicorn==21.2.0
